requirements.txt
flask==3.0.2
flask-cors==4.0.0
openai==1.12.0
python-dotenv==1.0.1 